+++ psutil/tests/test_bsd.py
@@ -16,8 +16,6 @@ import re
 import time
 import unittest
 
-import pytest
-
 import psutil
 from psutil import BSD
 from psutil import FREEBSD
@@ -26,6 +24,7 @@ from psutil import OPENBSD
 from psutil.tests import HAS_BATTERY
 from psutil.tests import TOLERANCE_SYS_MEM
 from psutil.tests import PsutilTestCase
+from psutil.tests import pytest
 from psutil.tests import retry_on_failure
 from psutil.tests import sh
 from psutil.tests import spawn_testproc