+++ psutil/tests/test_misc.py
@@ -18,8 +18,6 @@ import stat
 import sys
 import unittest
 
-import pytest
-
 import psutil
 import psutil.tests
 from psutil import POSIX
@@ -50,6 +48,7 @@ from psutil.tests import SCRIPTS_DIR
 from psutil.tests import PsutilTestCase
 from psutil.tests import mock
 from psutil.tests import process_namespace
+from psutil.tests import pytest
 from psutil.tests import reload_module
 from psutil.tests import sh
 from psutil.tests import system_namespace
@@ -634,26 +633,29 @@ class TestCommonModule(PsutilTestCase):
         else:
             from StringIO import StringIO
 
-        with redirect_stderr(StringIO()) as f:
-            debug("hello")
-            sys.stderr.flush()
+        with mock.patch.object(psutil._common, "PSUTIL_DEBUG", True):
+            with redirect_stderr(StringIO()) as f:
+                debug("hello")
+                sys.stderr.flush()
         msg = f.getvalue()
         assert msg.startswith("psutil-debug"), msg
         assert "hello" in msg
         assert __file__.replace('.pyc', '.py') in msg
 
         # supposed to use repr(exc)
-        with redirect_stderr(StringIO()) as f:
-            debug(ValueError("this is an error"))
+        with mock.patch.object(psutil._common, "PSUTIL_DEBUG", True):
+            with redirect_stderr(StringIO()) as f:
+                debug(ValueError("this is an error"))
         msg = f.getvalue()
         assert "ignoring ValueError" in msg
         assert "'this is an error'" in msg
 
         # supposed to use str(exc), because of extra info about file name
-        with redirect_stderr(StringIO()) as f:
-            exc = OSError(2, "no such file")
-            exc.filename = "/foo"
-            debug(exc)
+        with mock.patch.object(psutil._common, "PSUTIL_DEBUG", True):
+            with redirect_stderr(StringIO()) as f:
+                exc = OSError(2, "no such file")
+                exc.filename = "/foo"
+                debug(exc)
         msg = f.getvalue()
         assert "no such file" in msg
         assert "/foo" in msg